upModel

+++ src/main/java/org/example/ontap2/service/IUserService.java
@@ -4,11 +4,10 @@ import org.example.ontap2.model.User;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-
 @Service
 public interface IUserService {
     List<User> getAllUsers();
-    User getUserById(Integer id);
+    User getUserById(Long id);
     User saveUser(User user);
-    boolean deleteUser(Integer id);
+    boolean deleteUser(Long id);
 }
